Refactor raw-handler and paste-handler into their own files
