+++ packages/blocks/src/api/index.native.js
@@ -4,6 +4,7 @@ export {
 } from './factory';
 export {
 	default as parse,
+	getBlockAttributes,
 	parseWithAttributeSchema,
 } from './parser';
 export {
@@ -28,5 +29,5 @@ export {
 export {
 	isUnmodifiedDefaultBlock,
 } from './utils';
-export { getPhrasingContentSchema } from './raw-handling';
+export { pasteHandler, getPhrasingContentSchema } from './raw-handling';
 export { default as children } from './children';
