+++ packages/blocks/src/api/raw-handling/index.native.js
@@ -1,1 +1,3 @@
 export { getPhrasingContentSchema } from './phrasing-content';
+export { pasteHandler } from './paste-handler';
+export { rawHandler } from './raw-handler';
